// components/types.tsx
export enum TxResult {
  Success = 0,
  Failed = 1,
}
// ...
export interface Collection {
  collectionAddr: string;
  name: string;
  description: string;
  tokensCount: number;
  createdBy: CreatedBy;
  floorPrice: number;
}

interface CreatedBy {
  addr: string;
  name?: {
    name: string;
  };
}

// *Token Types*
export interface Tokens {
  tokens: {
    total: number;
    tokens: Token[];
  };
}

export interface Token {
  collectionAddr: string;
  name: string;
  price?: Price;
  priceExpiresAt?: string;
  saleType?: string;
  tokenId: string;
  highestCollectionBid: string;
  description: string;
  forSale: boolean;
  imageUrl: string;
  rarityOrder: number;
  traits: Trait[];
  id: string;
  lastSale: LastSale;
  highestCollectionBidEvent: {
    bidder: {
      addr: string;
    };
  };
}

interface LastSale {
  eventName: string;
  price: Price;
}

export interface Trait {
  name: string;
  value: string;
  rarityPercent: number;
}

interface Price {
  denom: string;
  amount: string;
}

// *Name Types*
export interface Names {
  names: {
    names: Name[];
  };
}

interface Name {
  name: string;
  records: Record[];
}

interface Record {
  id: string;
  recordName: string;
  recordValue: string;
  verified: boolean;
  createdAt: string;
  updatedAt: string;
  name: string;
  contractAddr: string;
}

// *Event Types*
export interface Events {
  events: {
    edges: Edge[];
  };
}

interface Edge {
  node: Node;
}

interface Node {
  eventName: string;
  price?: Price;
  data: Data;
  action?: string;
  id: string;
}